fix: split files that exceeded the max-lines ratchet
Move workspace-removal test fixtures and FakeGitBackend worktree helpers into dedicated modules so static CI stays under the 500-line cap.
